add invite system
User can invite new contacts by sending an email invite.
The new user receives an email via Mailgun with a link.
The link leads to a registration page with the inviting
users user-id appended to the url (this might be a security
vulnerability). When the new user registers, the inviting
users id is automatically added as a contact for the new user.